+++ Clients/TgDownloaderDesktop/Behaviors/NavigationViewHeaderMode.cs
@@ -1,0 +1,8 @@
+﻿namespace TgDownloaderDesktop.Behaviors;
+
+public enum NavigationViewHeaderMode
+{
+    Always,
+    Never,
+    Minimal
+}