+++ Clients/TgDownloaderDesktop/Contracts/Services/IActivationService.cs
@@ -1,0 +1,6 @@
+﻿namespace TgDownloaderDesktop.Contracts.Services;
+
+public interface IActivationService
+{
+    Task ActivateAsync(object activationArgs);
+}