+++ Clients/TgDownloaderDesktop/Contracts/Services/IPageService.cs
@@ -1,0 +1,6 @@
+﻿namespace TgDownloaderDesktop.Contracts.Services;
+
+public interface IPageService
+{
+    Type GetPageType(string key);
+}