+++ Clients/TgDownloaderDesktop/Contracts/ViewModels/INavigationAware.cs
@@ -1,0 +1,8 @@
+﻿namespace TgDownloaderDesktop.Contracts.ViewModels;
+
+public interface INavigationAware
+{
+    void OnNavigatedTo(object parameter);
+
+    void OnNavigatedFrom();
+}