+++ Clients/TgDownloaderDesktop/ViewModels/ContentGridViewModel.cs
@@ -1,0 +1,52 @@
+﻿using System.Collections.ObjectModel;
+using System.Windows.Input;
+
+using CommunityToolkit.Mvvm.ComponentModel;
+using CommunityToolkit.Mvvm.Input;
+
+using TgDownloaderDesktop.Contracts.Services;
+using TgDownloaderDesktop.Contracts.ViewModels;
+using TgDownloaderDesktop.Core.Contracts.Services;
+using TgDownloaderDesktop.Core.Models;
+
+namespace TgDownloaderDesktop.ViewModels;
+
+public partial class ContentGridViewModel : ObservableRecipient, INavigationAware
+{
+    private readonly INavigationService _navigationService;
+    private readonly ISampleDataService _sampleDataService;
+
+    public ObservableCollection<SampleOrder> Source { get; } = new ObservableCollection<SampleOrder>();
+
+    public ContentGridViewModel(INavigationService navigationService, ISampleDataService sampleDataService)
+    {
+        _navigationService = navigationService;
+        _sampleDataService = sampleDataService;
+    }
+
+    public async void OnNavigatedTo(object parameter)
+    {
+        Source.Clear();
+
+        // TODO: Replace with real data.
+        var data = await _sampleDataService.GetContentGridDataAsync();
+        foreach (var item in data)
+        {
+            Source.Add(item);
+        }
+    }
+
+    public void OnNavigatedFrom()
+    {
+    }
+
+    [RelayCommand]
+    private void OnItemClick(SampleOrder? clickedItem)
+    {
+        if (clickedItem != null)
+        {
+            _navigationService.SetListDataItemForNextConnectedAnimation(clickedItem);
+            _navigationService.NavigateTo(typeof(ContentGridDetailViewModel).FullName!, clickedItem.OrderID);
+        }
+    }
+}