+++ Clients/TgDownloaderDesktop/ViewModels/MainViewModel.cs
@@ -1,0 +1,10 @@
+﻿using CommunityToolkit.Mvvm.ComponentModel;
+
+namespace TgDownloaderDesktop.ViewModels;
+
+public partial class MainViewModel : ObservableRecipient
+{
+    public MainViewModel()
+    {
+    }
+}