+++ Clients/TgDownloaderDesktop/Views/DataGridPage.xaml.cs
@@ -1,0 +1,21 @@
+﻿using Microsoft.UI.Xaml.Controls;
+
+using TgDownloaderDesktop.ViewModels;
+
+namespace TgDownloaderDesktop.Views;
+
+// TODO: Change the grid as appropriate for your app. Adjust the column definitions on DataGridPage.xaml.
+// For more details, see the documentation at https://docs.microsoft.com/windows/communitytoolkit/controls/datagrid.
+public sealed partial class DataGridPage : Page
+{
+    public DataGridViewModel ViewModel
+    {
+        get;
+    }
+
+    public DataGridPage()
+    {
+        ViewModel = App.GetService<DataGridViewModel>();
+        InitializeComponent();
+    }
+}