+++ Clients/TgDownloaderDesktop/Views/ListDetailsPage.xaml.cs
@@ -1,0 +1,29 @@
+﻿using CommunityToolkit.WinUI.UI.Controls;
+
+using Microsoft.UI.Xaml.Controls;
+
+using TgDownloaderDesktop.ViewModels;
+
+namespace TgDownloaderDesktop.Views;
+
+public sealed partial class ListDetailsPage : Page
+{
+    public ListDetailsViewModel ViewModel
+    {
+        get;
+    }
+
+    public ListDetailsPage()
+    {
+        ViewModel = App.GetService<ListDetailsViewModel>();
+        InitializeComponent();
+    }
+
+    private void OnViewStateChanged(object sender, ListDetailsViewState e)
+    {
+        if (e == ListDetailsViewState.Both)
+        {
+            ViewModel.EnsureItemSelected();
+        }
+    }
+}