+++ Core/TgDownloaderDesktop.Core/Contracts/Services/ISampleDataService.cs
@@ -1,0 +1,13 @@
+﻿using TgDownloaderDesktop.Core.Models;
+
+namespace TgDownloaderDesktop.Core.Contracts.Services;
+
+// Remove this class once your pages/features are using your data.
+public interface ISampleDataService
+{
+    Task<IEnumerable<SampleOrder>> GetContentGridDataAsync();
+
+    Task<IEnumerable<SampleOrder>> GetGridDataAsync();
+
+    Task<IEnumerable<SampleOrder>> GetListDetailsDataAsync();
+}